requirements.txt
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
DateTime==5.5
fields==5.0.0
Flask==3.0.3
flask-marshmallow==1.2.1
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
jsonify==0.5
MarkupSafe==2.1.5
marshmallow==3.22.0
marshmallow-sqlalchemy==1.1.0
mysql-connector-python==9.0.0
packaging==24.1
pytz==2024.1
requests==2.32.3
setuptools==73.0.1
SQLAlchemy==2.0.32
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.4
zope.interface==7.0.1
